requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
FastAPI==0.104.1
uvicorn==0.24.0
httpx==0.25.2
